+++ src/Authenticator/Utils/MustHasUser.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CaliforniaMountainSnake\SimpleLaravelAuthSystem\Authenticator\Utils;
+
+use CaliforniaMountainSnake\SimpleLaravelAuthSystem\AuthUserEntity;
+
+trait MustHasUser
+{
+    /**
+     * @return AuthUserEntity|null
+     */
+    abstract public function getUserEntity(): ?AuthUserEntity;
+}