+++ src/TestUtils/RequireAdminRoles.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CaliforniaMountainSnake\SimpleLaravelAuthSystem\TestUtils;
+
+use CaliforniaMountainSnake\SimpleLaravelAuthSystem\Enums\AuthUserRoleEnum;
+
+trait RequireAdminRoles
+{
+    /**
+     * @return AuthUserRoleEnum[]
+     */
+    abstract public function getAdminRoles(): array;
+}